# docs/swagger.yaml
basePath: /api/v1
definitions:
  dto.AdvancePayload:
    properties:
      approved_amount:
        type: number
      comment:
        type: string
      currency:
        type: string
      requested_amount:
        type: number
    type: object
...
  dto.LoginRequest:
    properties:
      email:
        type: string
      password:
        type: string
    type: object
...
  dto.RefreshTokenRequest:
    properties:
      refreshToken:
        type: string
    type: object
  dto.RegisterRequest:
    properties:
      email:
        type: string
      full_name:
        type: string
      password:
        type: string
    type: object
  dto.StatusPayload:
    properties:
      comment:
        type: string
      status:
        type: string
    type: object
...
info:
  contact: {}
  description: REST API for the Business Trip and Expense Portal
  title: Travel Portal API
  version: "1.0"
paths:
...
  /audit/{entityType}/{entityId}/approvals:
    get:
      parameters:
      - description: Entity type
        enum:
        - trip_request
        - advance_request
        - expense_report
        in: path
        name: entityType
        required: true
        type: string
      - description: Entity ID
        in: path
        name: entityId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: List approval actions
      tags:
      - Audit
  /audit/{entityType}/{entityId}/logs:
    get:
      parameters:
      - description: Entity type
        enum:
        - trip_request
        - advance_request
        - expense_report
        in: path
        name: entityType
        required: true
        type: string
      - description: Entity ID
        in: path
        name: entityId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: List audit log entries
      tags:
      - Audit
  /auth/login:
    post:
      consumes:
      - application/json
      parameters:
      - description: Login payload
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/dto.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
      summary: Login via email and password
      tags:
      - Auth
  /auth/me:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get current user
      tags:
      - Auth
  /auth/refresh:
    post:
      consumes:
      - application/json
      parameters:
      - description: Refresh token payload
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/dto.RefreshTokenRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
      summary: Refresh access token
      tags:
      - Auth
  /auth/register:
    post:
      consumes:
      - application/json
      description: Creates an employee profile and issues default role.
      parameters:
      - description: Registration payload
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/dto.RegisterRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
      summary: Register new employee
      tags:
      - Auth
...
  /trip-requests/{tripId}/advance:
    get:
      parameters:
      - description: Trip ID
        in: path
        name: tripId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get advance for trip
      tags:
      - Advances
    post:
      consumes:
      - application/json
      parameters:
      - description: Trip ID
        in: path
        name: tripId
        required: true
        type: string
      - description: Advance payload
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/dto.AdvancePayload'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Create or update draft advance
      tags:
      - Advances
  /trip-requests/{tripId}/advance/status:
    patch:
      consumes:
      - application/json
      parameters:
      - description: Trip ID
        in: path
        name: tripId
        required: true
        type: string
      - description: Status payload
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/dto.StatusPayload'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Change advance status
      tags:
      - Advances
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"